refactor(backend): split variants/service.py into sub-package
Decomposes the 824-LOC variants/service.py into a sub-package at the
same import path:

  variants/service/
    __init__.py       (78)   re-exports every symbol the regression tests use
    errors.py         (26)   VEPError hierarchy (5 classes)
    validators.py     (143)  is_cnv_variant, validate_variant_id, _format_variant_for_vep
    cache_ops.py      (204)  variant_annotations table read/write helpers
    vep_api.py        (260)  Ensembl VEP REST client + response parser
    api.py            (142)  get_variant_annotation, get_variant_annotations_batch

Every file is under 500 LOC. Public import path unchanged:

    from app.variants.service import (
        get_variant_annotation,
        get_variant_annotations_batch,
        VEPError,
    )

Private helpers that the regression suites reach into are also
re-exported at the package level:

    from app.variants.service import (
        is_cnv_variant,
        validate_variant_id,
        _format_variant_for_vep,
        _parse_vep_response,
    )

tests/test_cnv_annotation.py, tests/test_variant_service_cnv.py, and
tests/test_variant_annotation_vep.py all keep working unchanged.

Counts unchanged: 907 passed, 9 skipped, 3 xfailed.
